move UnwrappedSequenceNumber to Utils
