+++ worker/include/RTC/SCTP/association/StreamResetHandler.hpp
@@ -4,7 +4,6 @@
 #include "common.hpp"
 #include "SharedInterface.hpp"
 #include "RTC/SCTP/association/TransmissionControlBlockInterface.hpp"
-#include "RTC/SCTP/common/UnwrappedSequenceNumber.hpp"
 #include "RTC/SCTP/packet/Packet.hpp"
 #include "RTC/SCTP/packet/chunks/ReConfigChunk.hpp"
 #include "RTC/SCTP/packet/parameters/IncomingSsnResetRequestParameter.hpp"
@@ -12,6 +11,7 @@
 #include "RTC/SCTP/packet/parameters/ReconfigurationResponseParameter.hpp"
 #include "RTC/SCTP/public/AssociationListener.hpp"
 #include "RTC/SCTP/tx/RetransmissionQueue.hpp"
+#include "Utils/UnwrappedSequenceNumber.hpp"
 #include "handles/BackoffTimerHandleInterface.hpp"
 #include <span>
 #include <vector>
@@ -162,7 +162,7 @@ namespace RTC
 			};
 
 		private:
-			using UnwrappedReConfigRequestSn = UnwrappedSequenceNumber<uint32_t>;
+			using UnwrappedReConfigRequestSn = Utils::UnwrappedSequenceNumber<uint32_t>;
 
 		public:
 			StreamResetHandler(
