+++ worker/meson.build
@@ -493,7 +493,6 @@ test_sources = [
   'test/src/RTC/SCTP/sctpCommon.cpp',
   'test/src/RTC/SCTP/association/TestStateCookie.cpp',
   'test/src/RTC/SCTP/association/TestNegotiatedCapabilities.cpp',
-  'test/src/RTC/SCTP/common/TestUnwrappedSequenceNumber.cpp',
   'test/src/RTC/SCTP/tx/TestOutstandingData.cpp',
   'test/src/RTC/SCTP/tx/TestRetransmissionErrorCounter.cpp',
   'test/src/RTC/SCTP/tx/TestRetransmissionQueue.cpp',
@@ -561,6 +560,7 @@ test_sources = [
   'test/src/Utils/TestNumber.cpp',
   'test/src/Utils/TestString.cpp',
   'test/src/Utils/TestTime.cpp',
+  'test/src/Utils/TestUnwrappedSequenceNumber.cpp',
   'mocks/src/MockShared.cpp',
   'mocks/src/handles/MockBackoffTimerHandle.cpp',
   'mocks/src/Channel/MockChannelMessageRegistrator.cpp',
